完成

+++ app/components/Modal.jsx
@@ -1,15 +1,30 @@
+import { useState } from 'react';
 import db from '../data/firebase';
 import styles from './Modal.module.css'
 import { doc, deleteDoc } from "firebase/firestore";
 
 const Modal = ({ selectedMovie, setSelectedMovie }) => {
+
+    const [password, setPassword] = useState('');
+    const [show, setShow] = useState(false);
+
     const handleModal = () => {
         setSelectedMovie(null);
     }
 
     const handleDelete = async () => {
-        await deleteDoc(doc(db, 'posts', selectedMovie.id));
-        setSelectedMovie(null);
+        if (password == 'js') {
+            await deleteDoc(doc(db, 'posts', selectedMovie.id));
+            setSelectedMovie(null);
+            setPassword('');
+        } else {
+            alert('パスワードが違います。');
+            return;
+        }
+    }
+
+    const showPass = () => {
+        setShow(true);
     }
 
     return (
@@ -23,11 +38,19 @@ const Modal = ({ selectedMovie, setSelectedMovie }) => {
                         ))}
                         <p className={styles.story}>{selectedMovie.story}</p>
                         <button className={styles.close} onClick={handleModal}>閉じる</button>
-                        <button className={styles.delete} onClick={handleDelete}>削除</button>
+                        <button className={styles.delete} onClick={showPass}>削除</button>
+                        {show && <div className={styles.password}>
+                            <p>本当に削除しますか？</p>
+                            <input autoComplete='off' placeholder='パスワードを入力してください。' value={password} onChange={(e) => setPassword(e.target.value)} type="text" />
+                            <button onClick={() => setShow(false)} className={`${styles.passBtns} ${styles.passclose}`}>閉じる</button>
+                            <button onClick={handleDelete} className={`${styles.passBtns} ${styles.del}`}>削除</button>
+                        </div>
+                        }
                     </div>
                 </div>
-            )}
-        </div>
+            )
+            }
+        </div >
     )
 }
